setup.py
"""Build script for megatext C++ extensions."""

from pybind11.setup_helpers import Pybind11Extension, build_ext
from setuptools import setup

ext_modules = [
    Pybind11Extension(
        "megatext.data._helpers",
        ["src/megatext/data/_helpers.cpp"],
        cxx_std=17,
    ),
]

setup(
    ext_modules=ext_modules,
    cmdclass={"build_ext": build_ext},
)
