pyproject.toml
[project]
name = "megatext"
version = "0.1.0"
description = "Simple, performant and scalable Jax LLM pretraining"
requires-python = ">=3.12"
dependencies = [
    # Core ML
    "jax>=0.8.1",
    "jaxlib>=0.8.1",
    "flax>=0.12.6",
    "optax>=0.2.6",
    "orbax-checkpoint>=0.11.33",
    # Data
    "array-record>=0.8.3",
    "grain>=0.2.15",
    "numpy",
    # Tokenizers & HF
    "transformers>=4.57",
    "sentencepiece>=0.2.1",
    "tiktoken>=0.12.0",
    # Checkpoint conversion
    "torch",
    "safetensors",
    "ml-dtypes",
    # Config & Logging
    "pydantic>=2.0",
    "omegaconf>=2.3.0",
    "absl-py>=2.3.1",
    "wandb",
    "tensorboardx>=2.6.4",
    # Google Cloud
    "gcsfs",
    "google-cloud-storage",
    # TPU
    "libtpu>=0.0.30; platform_machine == 'x86_64' and sys_platform == 'linux'",
    "google-tunix>=0.1.3",
    "pathwaysutils>=0.1.4",
    "cloud-tpu-diagnostics>=0.1.5",
    "cloud-accelerator-diagnostics>=0.1.1",
    "google-cloud-mldiagnostics>=0.5.10",
    "ml-goodput-measurement>=0.0.15",
    # Quantization & Kernels
    "aqtp>=0.9.0",
    "tokamax>=0.0.8",
    "qwix>=0.1.4",
    # Utilities
    "etils>=1.13.0",
    "jaxtyping>=0.3.3",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sympy>=1.14.0",
    "protobuf>=5.0",
    "pillow>=12.0",
    "uvloop>=0.19.0",
    "psutil",
    "packaging",
    "tpustat",
    # Dev & Test
    "pytest",
    "pytest-xdist",
]

[build-system]
requires = ["setuptools>=75", "pybind11>=2.13"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
megatext = ["**/*.yaml", "**/*.jinja2", "assets/**/*"]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.pytest.ini_options]
testpaths = ["tests"]
